Code cleanup part 1 - automatic

+++ src/rice/p2p/glacier/FragmentKeyFactory.java
@@ -35,15 +35,18 @@ advised of the possibility of such damage.
 
 *******************************************************************************/ 
 package rice.p2p.glacier;
-import java.util.*;
-import java.util.StringTokenizer;
 
 import rice.environment.random.RandomSource;
-import rice.p2p.commonapi.*;
+import rice.p2p.commonapi.IdFactory;
+import rice.p2p.commonapi.IdRange;
+import rice.p2p.commonapi.IdSet;
+import rice.p2p.commonapi.NodeHandleSet;
 import rice.p2p.multiring.MultiringIdFactory;
 import rice.p2p.multiring.RingId;
-import rice.pastry.Id;
+
+import java.util.Random;
 import java.util.SortedMap;
+import java.util.StringTokenizer;
 
 /**
  * DESCRIBE THE CLASS
@@ -127,7 +130,7 @@ public class FragmentKeyFactory implements IdFactory {
     String fragmentIdS = stok.nextToken();
     RingId key = FACTORY.buildRingId(rice.pastry.Id.build(keyRingS), rice.pastry.Id.build(keyNodeS));
 
-    return new FragmentKey(new VersionKey(key, Long.valueOf(versionS).longValue()), Integer.valueOf(fragmentIdS).intValue());
+    return new FragmentKey(new VersionKey(key, Long.valueOf(versionS)), Integer.valueOf(fragmentIdS));
   }
 
   public rice.p2p.commonapi.Id buildIdFromToString(char[] chars, int offset, int length) {
